omnistudio/requirements.txt
# --- Workflow (LangGraph) ---
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
langchain-core>=0.3.0
langchain-openai>=0.3.0

# --- Serveur (FastAPI/DSFR) ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
sse-starlette>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
slowapi>=0.1.9

# --- Utilitaires ---
jinja2>=3.1.0
openpyxl>=3.1.0
python-slugify>=8.0.0
ffmpeg-python
python-docx>=1.1.0
pdfplumber>=0.11.0
